# Rewritten parsers - ~3x performance increase