--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,4 @@ logs
 *.log
 coverage
 node_modules
+.idea