--- a/lib/javascript.js
+++ b/lib/javascript.js
@@ -33,11 +33,12 @@ function parseSimpleString(parser) {
  * @param parser
  * @param start
  * @param end
+ * @param noBuffer
  * @returns {*}
  */
-function convertBufferRange(parser, start, end) {
+function convertBufferRange(parser, start, end, noBuffer) {
 	// If returnBuffers is active, all return values are returned as buffers besides numbers and errors
-	if (parser.optionReturnBuffers === true) {
+	if (!noBuffer && parser.optionReturnBuffers === true) {
 		return parser.buffer.slice(start, end);
 	}
 
@@ -48,9 +49,10 @@ function convertBufferRange(parser, start, end) {
  * Parse a '+' redis simple string response but forward the offsets
  * onto convertBufferRange to generate a string.
  * @param parser
+ * @param noBuffer
  * @returns {*}
  */
-function parseSimpleStringViaOffset(parser) {
+function parseSimpleStringViaOffset(parser, noBuffer) {
 	var start = parser.offset;
 	var offset = parser.offset;
 	var length = parser.buffer.length;
@@ -61,7 +63,7 @@ function parseSimpleStringViaOffset(parser) {
 			var c2 = parser.buffer[offset++];
 			if (c2 === 10) { // \n
 				parser.offset = offset;
-				return convertBufferRange(parser, start, offset - 2);
+				return convertBufferRange(parser, start, offset - 2, noBuffer);
 			}
 		}
 	}
@@ -74,7 +76,14 @@ function parseSimpleStringViaOffset(parser) {
  * @returns {*}
  */
 function parseLength(parser) {
-	var string = parseSimpleString(parser);
+	var string;
+	/* istanbul ignore if  */
+	if (parser.buffer.length > 4096) {
+		 string = parseSimpleStringViaOffset(parser, true);
+	} else {
+		string = parseSimpleString(parser);
+	}
+	
 	if (string !== undefined) {
 		var length = +string;
 		if (length === -1) {