--- a/package.json
+++ b/package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha",
+    "benchmark": "node ./benchmark",
     "posttest": "jshint . && npm run coverage && npm run coverage:check",
     "coverage": "node ./node_modules/istanbul/lib/cli.js cover --preserve-comments ./node_modules/mocha/bin/_mocha -- -R spec",
     "coverage:check": "node ./node_modules/istanbul/lib/cli.js check-coverage --branch 100 --statement 100"
@@ -28,10 +29,12 @@
     "node": ">=0.10.0"
   },
   "devDependencies": {
+    "benchmark": "^2.1.0",
     "codeclimate-test-reporter": "^0.1.1",
     "intercept-stdout": "^0.1.2",
     "istanbul": "^0.4.0",
     "jshint": "^2.8.0",
+    "microtime": "^2.1.1",
     "mocha": "^2.3.2"
   },
   "optionalDependency": {