--- a/package.json
+++ b/package.json
@@ -30,11 +30,10 @@
   },
   "devDependencies": {
     "benchmark": "^2.1.0",
-    "codeclimate-test-reporter": "^0.1.1",
+    "codeclimate-test-reporter": "^0.3.1",
     "intercept-stdout": "^0.1.2",
     "istanbul": "^0.4.0",
     "jshint": "^2.8.0",
-    "microtime": "^2.1.1",
     "mocha": "^2.3.2"
   },
   "optionalDependency": {