Re-wrote the javascript parser, averages around 14k ops/s more and ~2-3x faster parsing.

Thoughts?

## Benchmarks
![image](https://cloud.githubusercontent.com/assets/5347038/15053515/ee7b08a2-12fa-11e6-80eb-69858e749cdf.png)

Also slightly tweaked the hiredis parser also, ~1-2% change, not really noticeable but every little helps

@luin @BridgeAR tagging you both in.