@BridgeAR 2016-05-05
@Salakar nice! But the optimization for big chunks is gone.