@Salakar 2016-05-05
ooo ye, let me look into that, woops. Guess it needs some kind of size check also to vary between chunking or standard?