@BridgeAR 2016-05-05
Yepp. If a bulk reply is bigger than the current chunk cache all following chunks until the bulk reply is complete and concat all cached chunks.