@Salakar 2016-05-08
Maybe we could persuade either @mscdex or @andrasq to work their performance wizardry / give this a once over look at - v2.0 branch. @ both - this is the main redis resp parser that the 2 main node redis clients for node use, based on:  http://redis.io/topics/protocol