@mscdex 2016-05-09
I decided to take a crack at this this evening (writing one from scratch) for fun. So far the results are promising, but I am far from finished yet. There are some more optimizations to be had and type implementations to be completed. Example results as the code stands right now though:

```
multiple chunks in a bulk string
================================
old x 50,943 ops/sec ±0.36% (92 runs sampled)
hiredis x 1,075,063 ops/sec ±0.39% (95 runs sampled)
new x 460,955 ops/sec ±0.82% (94 runs sampled)
mscdex x 1,334,433 ops/sec ±1.63% (93 runs sampled)

Fastest is mscdex

4mb bulk string
===============
old x 329 ops/sec ±0.47% (83 runs sampled)
hiredis x 469 ops/sec ±1.05% (79 runs sampled)
new x 327 ops/sec ±0.48% (65 runs sampled)
mscdex x 1,742 ops/sec ±1.30% (85 runs sampled)

Fastest is mscdex

simple string
=============
old x 2,920,941 ops/sec ±0.90% (93 runs sampled)
hiredis x 2,283,245 ops/sec ±0.44% (92 runs sampled)
new x 4,734,593 ops/sec ±0.37% (93 runs sampled)
mscdex x 5,090,149 ops/sec ±0.42% (91 runs sampled)

Fastest is mscdex

integer
=======
old x 2,688,599 ops/sec ±0.69% (88 runs sampled)
hiredis x 2,185,333 ops/sec ±0.24% (94 runs sampled)
new x 6,014,210 ops/sec ±0.48% (92 runs sampled)
mscdex x 19,786,130 ops/sec ±0.38% (92 runs sampled)

Fastest is mscdex
```

Some initial comments:

  * I suspect that the simple string parsing speed will be about the same as the "new" parser because IIRC most of the time is spent converting Buffers to strings, and there is not much that can be done about that outside of node core.

  * I removed the try-catch from the hiredis js interface since hiredis *could* be modified to not throw also, so removing the try-catch makes things more realistic IMHO when comparing benchmark results. This removal is reflected in the results above.

  * I found a bug with `startBigBuffer`. It is appending `lorem` which causes all of the parsers to fail (due to incorrect string length). Removing the `+ lorem` at the end fixes this.