@Salakar 2016-05-09
@mscdex this looks impressive! Be interesting to see how you achieved this

As for hiredis - I think we were deprecating support as in most cases js is way faster than native

Also, any thoughts on @luin 's big array perf on the [v2 proposal pr](https://github.com/NodeRedis/node-redis-parser/pull/3)