@mscdex 2016-05-09
@Salakar I actually pulled from the `v.2.0` branch, so the "new" parser is the one from #3 as of this writing.

I have not worked on array, error, or null implementations at all yet.