@andrasq 2016-05-10
@Salakar I only got a chance to look at this tonight, wrote an alternate number parser,
then got derailed when I tested with 2^64 - 1.

My approach was to gather the digits numerically into two numeric vars n and m, and return
either the numeric value of n if no overflow, or the string concatenation of n and m otherwise.
@mscdex may have done the same, my results are similar to his (once normalized)

However, when I tested this approach I found that parsing a big number causes the throughput
of all the other test runs to drop by a large amount (close to half), then spent the rest of the time
chasing that glitch.  Never did figure out how to avoid it.

```
old: + integer x 1,882,301 ops/sec ±0.57% (90 runs sampled)
hiredis: + integer x 1,289,520 ops/sec ±0.15% (95 runs sampled)
new: + integer x 3,270,845 ops/sec ±0.80% (90 runs sampled)
andras: + integer x 11,643,973 ops/sec ±0.52% (94 runs sampled)
```

My code, if you want to play with it
```
// gather digits numerically, return string if Number overflow
// 15 digits will not cause overflow, 16 can
function parseSimpleNumbers (parser) {
  var buf = parser.buffer, offset = parser.offset, length = buf.length
  var v, n = 0, m = 1, suffix = '', sign = ''

  if (buf[offset] === 0x2d) { sign = '-'; offset++ }
  var start = offset

  while (offset < length) {
    if (buf[offset] >= 0x30 && buf[offset] <= 0x39) { // '0'..'9'
      if (offset - start < 15) n = (n * 10) + buf[offset++] - 0x30
      else m = (m * 10) + buf[offset++] - 0x30
    }
    else if (buf[offset++] === 0x0a) break; // '\n'
  }
  parser.offset = offset
  if (offset - start > 15 && m !== 1) return sign + n + ('' + m).slice(1)
  return sign ? -n : n
}
```