@Salakar 2016-05-10
@andrasq that's an interesting way of doing that, incredibly performant, the benchmarks look identical to what @mscdex so does look like it's the same. Would never of thought of doing it that way, learn something new everyday!

There was two unused vars, removing those ups it by 2mil ops/s here.

Benchmark on my machine:
![image](https://cloud.githubusercontent.com/assets/5347038/15142712/3cebf78e-169e-11e6-90a4-0b4af760c1e8.png)