@andrasq 2016-05-10
nice!

whoops, the `v` and `suffix` were left over from experiments.
The glitch on my machine is there with both node 5 and 6 (though
node 6 is less affected):
```
node-v5.10.1:
andras: + integer x 11,653,357 ops/sec ±0.43% (87 runs sampled)
andras: + largeInteger x 1,347,323 ops/sec ±0.84% (93 runs sampled)
andras: + integer x 6,224,194 ops/sec ±0.82% (91 runs sampled)

node-v6.0.0:
andras: + integer x 12,459,959 ops/sec ±0.33% (89 runs sampled)
andras: + largeInteger x 1,516,039 ops/sec ±1.75% (90 runs sampled)
andras: + integer x 8,359,377 ops/sec ±0.88% (91 runs sampled)
```

the large integer used was
`var largeIntegerBuffer = new Buffer(':18446744073709551615\r\n');   // (2 ^ 64) - 1`