@Salakar 2016-05-10
@andrasq oh ok this happens here for me too:

```text
andras: + integer x 19,008,076 ops/sec ±1.24% (80 runs sampled)
andras: + largeInteger x 2,486,108 ops/sec ±1.02% (81 runs sampled)
andras: + integer x 15,435,560 ops/sec ±1.07% (84 runs sampled)
```

Weirdly only affects integers, if you string then large int then back to string it remains consistent:
```text
andras: + string x 4,615,010 ops/sec ±1.55% (87 runs sampled)
andras: + largeInteger x 2,527,927 ops/sec ±1.03% (87 runs sampled)
andras: + string x 4,612,367 ops/sec ±1.25% (85 runs sampled)
```