@andrasq 2016-05-10
it's weird.  Last night I tried `node --trace-gc` and the parse of integerBuffer shows
zero garbage collection before the largeInteger, lots of gc during largeInteger (as expected,
all those strings), but then lots of gc scavenging after largeInteger too even when back to
parsing integerBuffer again.  50-60 gc scavenge runs / second (every 15-20ms) and a couple
of mar-and-sweeps too.

it's possible the benchmark fits in the cpu cache without garbage collection, but once gc kicks in
gc will be a cache buster, with the cache contents bouncing between gc and the script, slowing it. (Which would suggest that the lower value is the more realistic figure to expect in production)
Still not bad though.  Good work that you're doing!

`qtimeit` is a node port of my php tool that I used for years before I started coding node.
I write lots of micro benchmarks, and qtimeit is great for one-liners.