@Salakar 2016-05-10
@andrasq could the issue be related to https://github.com/nodejs/node/issues/1671 I wonder, looks GC related