@andrasq 2016-05-10
@Salakar I wouldn't think so, that issue is specific to gc-ing Buffers.  Also, the same test run
behaves differently before and after an intervening largeInteger, so it's not gc specifically but
more the state of the system.